split-files: move more code
