+++ hashes/zkevm-keccak/src/keccak_circuit.rs
@@ -3,20 +3,23 @@ pub mod keccak_packed_multi;
 mod keccak_table;
 mod param;
 mod table;
-pub mod util;
-
 #[cfg(test)]
 mod test;
+pub mod util;
 
 use keccak_table::KeccakTable;
+use std::marker::PhantomData;
 
+use self::{cell_manager::*, keccak_packed_multi::*, param::*, table::*};
+use param::{
+    CHI_BASE_LOOKUP_TABLE, NUM_BYTES_PER_WORD, NUM_ROUNDS, NUM_WORDS_TO_ABSORB,
+    NUM_WORDS_TO_SQUEEZE, RHO_MATRIX,
+};
 use util::{
     constraint_builder::BaseConstraintBuilder,
     eth_types::Field,
     expression::{and, not, select, Expr},
     get_absorb_positions, get_num_bits_per_lookup, rotate, scatter, target_part_sizes,
-    CHI_BASE_LOOKUP_TABLE, NUM_BYTES_PER_WORD, NUM_ROUNDS, NUM_WORDS_TO_ABSORB,
-    NUM_WORDS_TO_SQUEEZE, RHO_MATRIX,
 };
 
 use crate::halo2_proofs::{
@@ -30,9 +33,6 @@ use crate::halo2_proofs::{
 use halo2_base::halo2_proofs::{circuit::AssignedCell, plonk::Assigned};
 use itertools::Itertools;
 use log::info;
-use std::marker::PhantomData;
-
-use self::{cell_manager::*, keccak_packed_multi::*, param::*, table::*};
 
 #[cfg(feature = "halo2-axiom")]
 type KeccakAssignedValue<'v, F> = AssignedCell<&'v Assigned<F>, F>;
